+++ apps/www/app/layout.tsx
@@ -32,6 +32,7 @@ export default function Layout({ children }: { children: ReactNode }) {
 			lang="en"
 			className={`${inter.className} ${jetbrainsMono.variable}`}
 			suppressHydrationWarning
+			data-scroll-behavior="smooth"
 		>
 			<body className="flex flex-col min-h-screen">
 				<RootProvider