+++ apps/www/middleware.ts
@@ -15,5 +15,5 @@ export function middleware(_request: NextRequest) {
 }
 
 export const config = {
-	matcher: "/:path*",
+	matcher: "/((?!monitoring).*)",
 };